Create README.md
My name is Maria Fernanda, but Fern suits me better. In all my studies, jobs and practices I have work to help other access places they found difficult to reach and experience,  that is why in this era of technology I wanna keep doing it by collaborating in the creation of quality useful products that make a difference.